firts commit

+++ src/main/java/com/etec/backend/controller/ForgotPasswordController.java
@@ -2,6 +2,7 @@ package com.etec.backend.controller;
 
 import com.etec.backend.dto.ChangePasswordRequestDTO;
 import com.etec.backend.dto.EmailRequestDTO;
+import com.etec.backend.dto.ResponseDTO;
 import com.etec.backend.entity.Auth;
 import com.etec.backend.entity.ForgotPassword;
 import com.etec.backend.repository.AuthRepository;
@@ -10,15 +11,13 @@ import com.etec.backend.utils.EmailService;
 
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.time.Instant;
 import java.util.Date;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Random;
 
 @RestController
@@ -31,59 +30,72 @@ public class ForgotPasswordController {
     private final ForgotPasswordRepository forgotPasswordRepository;
     private final PasswordEncoder passwordEncoder;
 
-    // enviar email para verificação
+    // Enviar email para verificação
     @PostMapping("/verifyMail/{email}")
-    public ResponseEntity<String> verifyEmail(@PathVariable String email) {
-        Auth auth = authRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("Por favor, forneça um email válido: " + email));
+    public ResponseDTO verifyEmail(@PathVariable String email) {
+        Optional<Auth> authOptional = authRepository.findByEmail(email);
 
+        if (authOptional.isEmpty()) {
+            return new ResponseDTO("Por favor, forneça um email válido: " + email );
+        }
+
+        Auth auth = authOptional.get();
         Long otp = otpGenerator();
         EmailRequestDTO emailRequestDTO = EmailRequestDTO.builder()
                 .to(email)
-                .from(email)
+                .from("heartcaresoftwares@gmail.com")
                 .text("Este é o OTP para sua solicitação de redefinição de senha: " + otp)
                 .subject("OTP para solicitação de redefinição de senha")
                 .build();
 
         ForgotPassword fp = ForgotPassword.builder()
                 .otp(otp)
-                .expirationTime(new Date(System.currentTimeMillis() + 1 * 60 * 1000)) // 4 minutos
+                .expirationTime(new Date(System.currentTimeMillis() + 1 * 60 * 1000)) // 1 minuto
                 .auth(auth)
                 .build();
 
         emailService.sendSimpleMessage(emailRequestDTO);
         forgotPasswordRepository.save(fp);
 
-        return ResponseEntity.ok("Email enviado para verificação!");
+        return new ResponseDTO("Email enviado para verificação!");
     }
 
     @PostMapping("/verifyOtp/{otp}/{email}")
-    public ResponseEntity<String> verifyOtp(@PathVariable Long otp, @PathVariable String email) {
-        Auth auth = authRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("Por favor, forneça um email válido!"));
+    public ResponseDTO verifyOtp(@PathVariable Long otp, @PathVariable String email) {
+        Optional<Auth> authOptional = authRepository.findByEmail(email);
 
-        ForgotPassword fp = forgotPasswordRepository.findByOtpAndUser(otp, auth)
-                .orElseThrow(() -> new RuntimeException("OTP inválido para o email: " + email));
+        if (authOptional.isEmpty()) {
+            return new ResponseDTO("Por favor, forneça um email válido!");
+        }
+
+        Auth auth = authOptional.get();
+        Optional<ForgotPassword> fpOptional = forgotPasswordRepository.findByOtpAndUser(otp, auth);
+
+        if (fpOptional.isEmpty()) {
+            return new ResponseDTO("OTP inválido para o email: " + email);
+        }
+
+        ForgotPassword fp = fpOptional.get();
 
         if (fp.getExpirationTime().before(Date.from(Instant.now()))) {
             forgotPasswordRepository.deleteById(fp.getId());
-            return new ResponseEntity<>("OTP expirou!", HttpStatus.EXPECTATION_FAILED);
+            return new ResponseDTO("OTP expirou!");
         }
 
-        return ResponseEntity.ok("OTP verificado!");
+        return new ResponseDTO("OTP verificado com sucesso!");
     }
 
     @PostMapping("/changePassword/{email}")
-    public ResponseEntity<String> changePasswordHandler(@RequestBody ChangePasswordRequestDTO changePasswordRequestDTO,
-            @PathVariable String email) {
+    public ResponseDTO changePasswordHandler(@RequestBody ChangePasswordRequestDTO changePasswordRequestDTO,
+                                             @PathVariable String email) {
         if (!Objects.equals(changePasswordRequestDTO.password(), changePasswordRequestDTO.repeatPassword())) {
-            return new ResponseEntity<>("Por favor, digite a senha novamente!", HttpStatus.EXPECTATION_FAILED);
+            return new ResponseDTO("Por favor, digite a senha novamente!");
         }
 
         String encodedPassword = passwordEncoder.encode(changePasswordRequestDTO.password());
         authRepository.updatePassword(email, encodedPassword);
 
-        return ResponseEntity.ok("Senha alterada com sucesso!");
+        return new ResponseDTO("Senha alterada com sucesso!");
     }
 
     private Long otpGenerator() {
